update: action, route, services, css, ui; refactor: public

+++ app/lib/session.js
@@ -30,11 +30,12 @@ export async function decrypt(session) {
         })
         return payload
     } catch (error) {
-        console.log('Failed to verify session')
+        console.error('Failed to verify session:', error.message)
+        throw error
     }
 }
 
-export async function createSession({ data }) {
+export async function createSession(data) {
     const { userId, username, email } = data
     if (
         !userId || typeof userId !== 'string'
@@ -44,28 +45,53 @@ export async function createSession({ data }) {
         throw new Error('Invalid session data');
     }
 
-    const expiresAt = new Date(Date.now() + 7 * 24 * 60 * 60 * 1000);
-    const session = await encrypt({ userId, username, email, expiresAt });
-    const cookieStore = await cookies();
+    try {
+        const expiresAt = new Date(Date.now() + 7 * 24 * 60 * 60 * 1000);
+        const session = await encrypt({ userId, username, email, expiresAt });
+        const cookieStore = await cookies();
 
-    cookieStore.set('session', session, {
-        httpOnly: true,
-        secure: process.env.NODE_ENV === 'production',
-        expires: expiresAt,
-        sameSite: 'lax',
-        path: '/',
-    })
+        cookieStore.set('session', session, {
+            httpOnly: true,
+            secure: process.env.NODE_ENV === 'production',
+            expires: expiresAt,
+            sameSite: 'lax',
+            path: '/',
+        })
+
+        return true
+    }
+    catch (err) {
+        console.error('Failed to create session:', err.message)
+        return false
+    }
 }
 
 export async function getSession() {
     const cookieStore = await cookies()
     const session = cookieStore.get('session')?.value
+    if (!session) return null
     return await decrypt(session)
 }
 
 export async function deleteSession() {
     const cookieStore = await cookies()
     const id = (await getSession())?.userId
-    await sql`delete from storage.session where id = ${id}`
-    cookieStore.delete('session')
+
+    try {
+        if (!id) return;
+
+        const res = await sql`delete from storage.session where id = ${id}`
+
+        if (res.count === 0) {
+            console.warn('No session found to delete for user ID:', id)
+            return false
+        }
+
+        cookieStore.delete('session')
+
+        return true
+    } catch (error) {
+        console.error('Failed to delete session:', error.message)
+        return false
+    }
 }
